#671-refactor-user-components
